// src/interfaces/ebooks.interface.ts
export interface IBookRequest {
  author: string;
  name: string;
  category: string;
  dateRelease: string;
  bookCover: string;
  value: number;
  description: string;
  publishingCompany: string;
  language: string;
  editionNumber: string;
  numberPages: string;
  country: string;
  isbn: string;
}